feat: enable invitely authentication
- Switch to FastAPI app with lazy initialization for Lambda
- Remove unused handler.py
- Add Bearer auth security to openapi.json
- Update curl examples with Authorization header
