+++ app/Dockerfile
@@ -11,18 +11,21 @@ RUN pip install --no-cache-dir \
     openai>=1.0.0 \
     presidio-analyzer>=2.2 \
     presidio-anonymizer>=2.2 \
+    presidio-image-redactor>=0.0.53 \
+    pillow>=10.0.0 \
     spacy>=3.7 \
     spacy-llm>=0.7.3 \
     langchain>=0.3.27 \
     langchain-community>=0.3.27 \
     confection>=0.1.3 \
-    jinja2
+    jinja2 \
+    scalar-fastapi>=1.6.0
 
 # Download spaCy English model
 RUN python -m spacy download en_core_web_sm
 
 # Copy application code
-COPY app/handler.py app/config.cfg ${LAMBDA_TASK_ROOT}/
+COPY app/app.py app/auth.py app/config.cfg ${LAMBDA_TASK_ROOT}/
 
-# Set handler
-CMD ["handler.handler"]
+# Set handler (Mangum wraps FastAPI)
+CMD ["app.handler"]
